pull code into only subclass
